pyproject.toml
[tool.poetry]
name = "dialog-keeper"
version = "0.1.0"
description = "Corporate Communications Gateway for Telegram with AI-powered RAG"
authors = ["Dialog Keeper Team"]
license = "GPL-3.0"
readme = "README.md"
repository = "https://github.com/xn-nil-x/dialog-keeper"
keywords = ["telegram", "bot", "ai", "rag", "customer-support"]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.7"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
openai = "^1.7.0"
anthropic = "^0.8.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
qdrant-client = "^1.7.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
celery = "^5.3.4"
httpx = "^0.26.0"
aiohttp = "^3.9.1"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
structlog = "^23.3.0"
tenacity = "^8.2.3"
sentry-sdk = "^1.39.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
black = "^23.12.1"
mypy = "^1.8.0"
ruff = "^0.1.9"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

